market_research/requirements.txt
requests
PyYAML
openai
panda